+++ CHANGELOG.md
@@ -24,3 +24,4 @@ The following changes have been completed, and will be included in the next tagg
 ### New
 
 - added `roundDown()` function
+- added `roundUp()` function